Removed unnecessary code.
I remove a 'char *ptr;' and some related code that was doing
nothing. Seems it was related to some old debugging. I have found that
the comments in conmsg.c are very confusing. I don't know what they
mean (2 different sets). One set appear to be for some debugging for
something that needs to be fixed.