Begun adding IPV6 support, reformatted code and added a few FIXME comments.
The existing network code was out of date so I began adding updated
networking code and support for IPV6. While it compiles it may not yet
be working. Some reformatting has been done and there are several
FIXME sections where the code is confusing and inconsistent. This code
needs further refactoring.